requirements-local.txt
--index-url https://download.pytorch.org/whl/cpu
torch
transformers==4.48.0
datasets==2.19.0
openai==1.40.0
flask==3.0.0
requests==2.31.0
pyserini==0.22.1
sentence-transformers==2.7.0
faiss-cpu==1.7.4
ipython==8.26.0
jupyterlab==4.2.4
python-dotenv==1.0.1
PyYAML==6.0.2
tqdm==4.66.5
